test cases: diff_tests: fix type of aliased symbol
This was never intended to be different and such discrepancies are now
detected by Clang resulting in warnings.

PiperOrigin-RevId: 930423155

+++ test_cases/diff_tests/symbol/alias.1.c
@@ -4,4 +4,4 @@ extern int z __attribute__ ((weak, alias ("x")));
 
 int a() { return 0; }
 int b() { return 1; }
-extern long c() __attribute__ ((alias ("b")));
+extern int c() __attribute__ ((alias ("b")));
